// codegen/crm/associations/v4/models/BatchInputPublicAssociationMultiArchive.ts
/**
 * CRM Associations
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { PublicAssociationMultiArchive } from '../models/PublicAssociationMultiArchive';

export class BatchInputPublicAssociationMultiArchive {
    'inputs': Array<PublicAssociationMultiArchive>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "inputs",
            "baseName": "inputs",
            "type": "Array<PublicAssociationMultiArchive>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BatchInputPublicAssociationMultiArchive.attributeTypeMap;
    }

    public constructor() {
    }
}