// codegen/crm/associations/v4/models/BatchInputPublicFetchAssociationsBatchRequest.ts
/**
 * Associations
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { PublicFetchAssociationsBatchRequest } from '../models/PublicFetchAssociationsBatchRequest';

export class BatchInputPublicFetchAssociationsBatchRequest {
    'inputs': Array<PublicFetchAssociationsBatchRequest>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "inputs",
            "baseName": "inputs",
            "type": "Array<PublicFetchAssociationsBatchRequest>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BatchInputPublicFetchAssociationsBatchRequest.attributeTypeMap;
    }

    public constructor() {
    }
}